producer/requirements.txt
kafka-python
pyvmomi